+++ src/components/pages/HomePage/index.js
@@ -1,12 +1,12 @@
 import React from "react";
 import { useSelector } from "react-redux";
-import LoginedHome from "./LoginedHome";
-import NotLoginedHome from "./NotLoginedHome";
+import LoginedHome from "../LoginedHome";
+import SignInPage from "../SignInPage";
 
 const Home = () => {
-  const session = useSelector((state) => state.session);
+  const { isLogined } = useSelector((state) => state.session);
 
-  return session.isLogined ? <LoginedHome /> : <NotLoginedHome />;
+  return isLogined ? <LoginedHome /> : <SignInPage />;
 };
 
 export default Home;
