+++ src/components/pages/UserPage/index.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import { useDispatch, useSelector } from "react-redux";
+import { useDispatch } from "react-redux";
 import { useLocation } from "react-router";
 import { endLoading, startLoading } from "../../../modules/loading";
 import { getPostsAPI } from "../../../utils/API";
